+++ packages/gitbook/src/components/AIChat/AIResponseFeedback.tsx
@@ -31,8 +31,8 @@ export function AIResponseFeedback(props: {
                 label={tString(language, 'was_this_helpful_positive_label')}
                 variant="blank"
                 className={tcls(
-                    'animate-fadeIn overflow-hidden text-tint-subtle transition-all',
-                    rating !== null && rating !== 1 && 'px-0 text-[0] opacity-0'
+                    'animate-fade-in overflow-hidden text-tint-subtle transition-all',
+                    rating !== null && rating !== 1 && 'px-0 text-[0rem] opacity-0'
                 )}
                 size="medium"
                 style={{ animationDuration: '.5s' }}
@@ -47,8 +47,8 @@ export function AIResponseFeedback(props: {
                 label={tString(language, 'was_this_helpful_negative_label')}
                 variant="blank"
                 className={tcls(
-                    'animate-fadeIn overflow-hidden text-tint-subtle transition-all',
-                    rating !== null && rating !== -1 && 'px-0 text-[0] opacity-0'
+                    'animate-fade-in overflow-hidden text-tint-subtle transition-all',
+                    rating !== null && rating !== -1 && 'px-0 text-[0rem] opacity-0'
                 )}
                 size="medium"
                 style={{ animationDelay: '.2s', animationDuration: '.5s' }}
@@ -59,8 +59,8 @@ export function AIResponseFeedback(props: {
             />
             {rating !== null ? (
                 <span
-                    className="ml-2 animate-fadeIn text-tint-subtle"
-                    style={{ animationDelay: '.3s', animationDuration: '.5s' }}
+                    className="ml-2 animate-fade-in-slow text-tint-subtle"
+                    style={{ animationDelay: '.3s' }}
                 >
                     {t(language, 'was_this_helpful_thank_you')}
                 </span>