+++ packages/gitbook/src/components/DocumentView/Hint.tsx
@@ -21,7 +21,7 @@ export function Hint(props: BlockProps<DocumentBlockHint>) {
                 'hint',
                 'transition-colors',
                 'rounded-md',
-                hasHeading ? 'rounded-l' : null,
+                hasHeading ? 'rounded-l-sm' : null,
                 'straight-corners:rounded-none',
                 'circular-corners:rounded-xl',
                 'overflow-hidden',
@@ -52,8 +52,10 @@ export function Hint(props: BlockProps<DocumentBlockHint>) {
             {hasHeading ? (
                 <Block
                     style={tcls(
-                        'flip-heading-hash !py-4 items-start px-4 pl-3 text-[1em] *:flex-none',
-                        hasHeading ? hintStyle.header : null
+                        'w-full items-start py-4! pl-3 text-[1em] *:flex-none',
+                        // Heading hash styles
+                        'flip-heading-hash pr-8',
+                        hintStyle.header
                     )}
                     ancestorBlocks={[...ancestorBlocks, block]}
                     {...contextProps}