+++ packages/gitbook/src/components/PageAside/PageAside.tsx
@@ -48,15 +48,15 @@ export function PageAside(props: {
                 'group/aside',
                 'hidden',
                 'xl:flex',
-                'chat-open:xl:max-3xl:hidden',
-                'chat-open:xl:max-3xl:opacity-0',
+                'xl:max-3xl:chat-open:hidden',
+                'xl:max-3xl:chat-open:opacity-0',
                 'max-w-56',
-                'chat-open:xl:max-3xl:*:w-56',
-                'chat-open:xl:max-3xl:max-w-0',
-                'chat-open:xl:max-3xl:ml-0',
+                'xl:max-3xl:*:chat-open:w-56',
+                'xl:max-3xl:chat-open:max-w-0',
+                'xl:max-3xl:chat-open:ml-0',
 
                 'motion-safe:xl:transition-all motion-safe:xl:duration-300',
-                'motion-safe:[transition-behavior:allow-discrete]',
+                'motion-safe:transition-discrete',
 
                 'flex-col',
                 'basis-56',
@@ -73,30 +73,31 @@ export function PageAside(props: {
                 'lg:max-h-screen',
 
                 // With header
-                'site-header:lg:top-16',
-                'site-header:lg:max-h-[calc(100vh_-_4rem)]',
+                'lg:site-header:top-16',
+                'lg:site-header:max-h-[calc(100vh-4rem)]',
 
                 // With header & sections
-                'site-header-sections:lg:top-[6.75rem]',
-                'site-header-sections:lg:max-h-[calc(100vh_-_6.75rem)]',
+                'lg:site-header-sections:top-27',
+                'lg:site-header-sections:max-h-[calc(100vh-6.75rem)]',
 
                 // When in api page mode, we display it as an overlay on non-large resolutions
-                'page-api-block:xl:max-2xl:z-10',
-                'page-api-block:xl:max-2xl:fixed',
-                'page-api-block:xl:max-2xl:right-8',
-                'page-api-block:xl:max-2xl:w-56',
-                'page-api-block:xl:max-2xl:bg-tint/9',
-                'page-api-block:xl:max-2xl:contrast-more:bg-tint',
-                'page-api-block:xl:max-2xl:backdrop-blur-lg',
-                'page-api-block:xl:max-2xl:border',
-                'page-api-block:xl:max-2xl:border-tint',
-                'page-api-block:xl:max-2xl:hover:shadow-lg',
-                'page-api-block:xl:max-2xl:hover:shadow-tint-12/1',
-                'page-api-block:xl:max-2xl:dark:hover:shadow-tint-1/1',
-                'page-api-block:xl:max-2xl:rounded-md',
-                'page-api-block:xl:max-2xl:h-auto',
-                'page-api-block:xl:max-2xl:my-4',
-                'page-api-block:p-2'
+                'xl:max-2xl:page-api-block:z-10',
+                'xl:max-2xl:page-api-block:fixed',
+                'xl:max-2xl:page-api-block:right-8',
+                'xl:max-2xl:page-api-block:w-56',
+                'xl:max-2xl:page-api-block:bg-tint/9',
+                'xl:max-2xl:contrast-more:page-api-block:bg-tint',
+                'xl:max-2xl:page-api-block:backdrop-blur-lg',
+                'xl:max-2xl:page-api-block:border',
+                'xl:max-2xl:page-api-block:border-tint',
+                'xl:max-2xl:page-api-block:hover:shadow-lg',
+                'xl:max-2xl:page-api-block:hover:shadow-tint-12/1',
+                'xl:max-2xl:dark:page-api-block:hover:shadow-tint-1/1',
+                'xl:max-2xl:page-api-block:rounded-md',
+                'xl:max-2xl:page-api-block:h-auto',
+                'xl:max-2xl:page-api-block:my-9',
+                'page-api-block:px-2',
+                'page-api-block:py-1.5'
             )}
         >
             {page.layout.outline ? (
@@ -104,7 +105,7 @@ export function PageAside(props: {
                     <div
                         className={tcls(
                             'hidden',
-                            'page-api-block:xl:max-2xl:flex',
+                            'xl:max-2xl:page-api-block:flex',
                             'text-xs',
                             'tracking-wide',
                             'font-semibold',
@@ -123,7 +124,7 @@ export function PageAside(props: {
                                 'size-3',
                                 'opacity-6',
                                 'ml-auto',
-                                'page-api-block:xl:max-2xl:group-hover/aside:hidden'
+                                'xl:max-2xl:page-api-block:group-hover/aside:hidden'
                             )}
                         />
                     </div>
@@ -139,16 +140,16 @@ export function PageAside(props: {
 
                             'sticky',
                             'lg:top:0',
-                            'site-header:lg:top-16',
-                            'site-header-sections:lg:top-[6.75rem]',
+                            'lg:site-header:top-16',
+                            'lg:site-header-sections:top-27',
 
                             'gap-6',
-                            'pt-8',
+                            'pt-10',
 
-                            'page-api-block:xl:max-2xl:py-0',
+                            'xl:max-2xl:page-api-block:py-0',
                             // Hide it for api page, until hovered
-                            'page-api-block:xl:max-2xl:hidden',
-                            'page-api-block:xl:max-2xl:group-hover/aside:flex'
+                            'xl:max-2xl:page-api-block:hidden',
+                            'xl:max-2xl:page-api-block:group-hover/aside:flex'
                         )}
                     >
                         {document ? (
@@ -165,15 +166,15 @@ export function PageAside(props: {
                                 'border-t',
                                 'first:border-none',
                                 'border-tint-subtle',
-                                'py-4',
+                                'py-6',
                                 'first:pt-0',
-                                'page-api-block:xl:max-2xl:px-3',
+                                'xl:max-2xl:page-api-block:px-3',
                                 'empty:hidden'
                             )}
                         >
                             {withPageFeedback ? (
                                 <React.Suspense fallback={null}>
-                                    <PageFeedbackForm pageId={page.id} className={tcls('mt-2')} />
+                                    <PageFeedbackForm pageId={page.id} />
                                 </React.Suspense>
                             ) : null}
                             {customization.git.showEditLink && space.gitSync?.url && page.git ? (
@@ -236,8 +237,8 @@ export function PageAside(props: {
             ) : null}
             <div
                 className={tcls(
-                    'sticky bottom-0 z-10 mt-auto flex flex-col bg-tint-base theme-gradient-tint:bg-gradient-tint theme-gradient:bg-gradient-primary theme-muted:bg-tint-subtle pb-4 page-api-block:xl:max-2xl:hidden page-api-block:xl:max-2xl:pb-0 page-api-block:xl:max-2xl:group-hover/aside:flex [html.sidebar-filled.theme-bold.tint_&]:bg-tint-subtle',
-                    'page-api-block:xl:max-2xl:bg-transparent'
+                    'sticky bottom-0 z-10 mt-auto flex flex-col bg-tint-base theme-gradient-tint:bg-gradient-tint theme-gradient:bg-gradient-primary theme-muted:bg-tint-subtle pb-4 xl:max-2xl:page-api-block:hidden xl:max-2xl:page-api-block:pb-0 xl:max-2xl:page-api-block:group-hover/aside:flex [html.sidebar-filled.theme-bold.tint_&]:bg-tint-subtle',
+                    'xl:max-2xl:page-api-block:bg-transparent'
                 )}
             >
                 {/* Mode Switcher */}