+++ packages/gitbook/src/components/TableOfContents/Trademark.tsx
@@ -22,7 +22,7 @@ export function Trademark(props: {
         <div
             className={tcls(
                 'relative',
-                'z-[2]',
+                'z-2',
                 'lg:absolute',
 
                 'left-0',
@@ -40,6 +40,10 @@ export function Trademark(props: {
                 '[html.sidebar-filled.theme-muted_&]:bg-tint-base',
                 '[html.sidebar-filled.theme-bold.tint_&]:bg-tint-base',
 
+                'rounded-lg',
+                'straight-corners:rounded-none',
+                'circular-corners:rounded-2xl',
+
                 'before:hidden',
                 'lg:before:block',
                 'before:content-[""]',
@@ -47,7 +51,7 @@ export function Trademark(props: {
                 'before:inset-x-0',
                 'before:bottom-full',
                 'before:h-8',
-                'before:bg-gradient-to-b',
+                'before:bg-linear-to-b',
                 'before:from-transparent',
                 'before:to-tint-base',
                 'sidebar-filled:before:to-tint-subtle',
@@ -55,7 +59,7 @@ export function Trademark(props: {
                 '[html.sidebar-filled.theme-bold.tint_&]:before:to-tint-subtle',
                 '[html.sidebar-filled.theme-muted_&]:before:to-tint-base',
                 '[html.sidebar-filled.theme-bold.tint_&]:before:to-tint-base',
-                'page-no-toc:before:!to-transparent'
+                'page-no-toc:before:to-transparent!'
             )}
         >
             <TrademarkLink {...props} />
@@ -85,9 +89,9 @@ export function TrademarkLink(props: {
             href={url.toString()}
             className={tcls(
                 'text-sm',
-                // 'page-no-toc:lg:max-xl:text-xs',
-                // 'page-no-toc:lg:max-xl:px-3',
-                // 'page-no-toc:lg:max-xl:py-3',
+                // 'lg:max-xl:page-no-toc:text-xs',
+                // 'lg:max-xl:page-no-toc:px-3',
+                // 'lg:max-xl:page-no-toc:py-3',
                 'font-semibold',
                 'text-tint',
 
@@ -98,7 +102,7 @@ export function TrademarkLink(props: {
                 'py-4',
 
                 'sidebar-filled:px-3',
-                'sidebar-filled:page-no-toc:lg:px-5',
+                'lg:sidebar-filled:page-no-toc:px-5',
 
                 'rounded-lg',
                 'straight-corners:rounded-none',
@@ -124,14 +128,14 @@ export function TrademarkLink(props: {
                 icon="gitbook"
                 className={tcls(
                     'size-5',
-                    // 'page-no-toc:lg:max-xl:size-4',
+                    // 'lg:max-xl:page-no-toc:size-4',
                     'shrink-0'
                 )}
             />
             <span
                 className={tcls(
                     'ml-3'
-                    // 'page-no-toc:lg:max-xl:ml-2'
+                    // 'lg:max-xl:page-no-toc:ml-2'
                 )}
             >
                 {t(language, 'powered_by_gitbook')}